pyproject.toml
[project]
name = "browsercontrol"
version = "0.1.0"
description = "MCP server for browser automation with Set of Marks (SoM) - AI agents can see and interact with web pages using numbered element IDs"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Aditya Sasidhar" }
]
keywords = ["mcp", "browser", "automation", "playwright", "ai", "agent", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: Browsers",
]
dependencies = [
    "fastmcp>=2.14.2",
    "playwright>=1.49.0",
    "markdownify>=0.14.1",
    "pillow>=11.0.0",
]

[project.urls]
Homepage = "https://github.com/adityasasidhar/browsercontrol"
Repository = "https://github.com/adityasasidhar/browsercontrol"

[project.scripts]
browsercontrol = "browsercontrol.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["browsercontrol"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
